energy production handled as float

+++ packages/modules/fronius/inverter.py
@@ -62,8 +62,8 @@ class FroniusInverter:
         if gen24:
             _, counter = self.__sim_count.sim_count(power, topic=topic, data=self.__simulation, prefix="pv")
         else:
-            counter = int(response.json()["Body"]["Data"]["Site"]["E_Total"])
-            daily_yield = int(response.json()["Body"]["Data"]["Site"]["E_Day"])
+            counter = float(response.json()["Body"]["Data"]["Site"]["E_Total"])
+            daily_yield = float(response.json()["Body"]["Data"]["Site"]["E_Day"])
             counter, counter_start, counter_offset = self.__calculate_offset(counter, daily_yield)
             counter = counter + counter2
             if counter > 0 and self.component_config["configuration"]["ip_address2"] == "none":
@@ -90,12 +90,12 @@ class FroniusInverter:
                                     params=params, timeout=3)
             response.raise_for_status()
             try:
-                power2 = int(response.json()["Body"]["Data"]["Site"]["P_PV"])
+                power2 = float(response.json()["Body"]["Data"]["Site"]["P_PV"])
             except TypeError:
                 # Ohne PV Produktion liefert der WR 'null', ersetze durch Zahl 0
                 power2 = 0
             if not self.component_config["configuration"]["gen24"]:
-                counter2 = int(response.json()["Body"]["Data"]["Site"]["E_Total"])
+                counter2 = float(response.json()["Body"]["Data"]["Site"]["E_Total"])
         else:
             power2 = 0
         return power2, counter2
@@ -105,13 +105,13 @@ class FroniusInverter:
         if ramdisk:
             try:
                 with open("/var/www/html/openWB/ramdisk/pvkwh_offset", "r") as f:
-                    counter_offset = int(f.read())
+                    counter_offset = float(f.read())
             except FileNotFoundError as e:
                 log.MainLogger().exception(str(e))
                 counter_offset = 0
             try:
                 with open("/var/www/html/openWB/ramdisk/pvkwh_start", "r") as f:
-                    counter_start = int(f.read())
+                    counter_start = float(f.read())
             except FileNotFoundError as e:
                 log.MainLogger().exception(str(e))
                 counter_start = counter - daily_yield
